담기버튼 UI구현

+++ app/products/models/product_option.py
@@ -92,7 +92,7 @@ class ProductOption(models.Model):
 
     @property
     def price_per_unit(self):
-        return self.price * self.unit
+        return int(self.price * self.unit)
 
 
 class ProductOptionUnit(models.Model):
